Split nodes and workflows in engine
