+++ nipype/__init__.py
@@ -83,6 +83,6 @@ except:
     pass
 
 
-from .pipeline import Node, MapNode, JoinNode, Workflow, ConditionalNode
+from .pipeline import engine
 from .interfaces import (DataGrabber, DataSink, SelectFiles,
                          IdentityInterface, Rename, Function, Select, Merge)
