+++ nipype/pipeline/__init__.py
@@ -7,4 +7,5 @@ Package contains modules for generating pipelines using interfaces
 
 from __future__ import absolute_import
 __docformat__ = 'restructuredtext'
-from .engine import Node, MapNode, JoinNode, Workflow, ConditionalNode
+
+from .engine import *
