+++ nipype/pipeline/engine/tests/test_join.py
