+++ setup.py
@@ -380,9 +380,10 @@ def main(**extra_args):
                     'nipype.interfaces.vista',
                     'nipype.interfaces.vista.tests',
                     'nipype.pipeline',
+                    'nipype.pipeline.engine',
+                    'nipype.pipeline.engine.tests',
                     'nipype.pipeline.plugins',
                     'nipype.pipeline.plugins.tests',
-                    'nipype.pipeline.tests',
                     'nipype.testing',
                     'nipype.testing.data',
                     'nipype.testing.data.bedpostxout',
